+++ tests/test_model_definition/test_model_definition.py
@@ -228,7 +228,7 @@ def test_binary_error_without_length_model_definition():
                 database = database
                 metadata = metadata
 
-            test: bytes = ormar.LargeBinary(primary_key=True)
+            test: bytes = ormar.LargeBinary(primary_key=True, max_length=-1)
 
 
 @typing.no_type_check
@@ -241,7 +241,7 @@ def test_string_error_in_model_definition():
                 database = database
                 metadata = metadata
 
-            test: str = ormar.String(primary_key=True)
+            test: str = ormar.String(primary_key=True, max_length=0)
 
 
 @typing.no_type_check